Add basic usdc contract interaction
