--- src/routes/strategy/ethlisbon/smart-contract.svelte
@@ -1,22 +1,0 @@
-<script>
-    import {TabContent, TabPane} from 'sveltestrap';
-    import StrategyNav from '$lib/strategy/StrategyNav.svelte';
-
-</script>
-
-<svelte:head>
-  <title>EthLisbon hackathon example</title>
-  <meta name="description" content="An example strategy">
-</svelte:head>
-
-
-<div class="row">
-  <div class="col-md-12">
-    <p>Address: <strong>0x6590D62f2Af8717295563514239A71B36FA78341</strong></p>
-
-    <p>
-      <a class="body-link" href="https://polygonscan.com/address/0x6590D62f2Af8717295563514239A71B36FA78341#code">View the smart contract on Polyscan</a>
-    </p>
-  </div>
-</div>
-
